# be/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouScan Dashboard API",
    "description": "Anonymous widget dashboards — create a dashboard, add line/bar/text widgets, restore later by key. The frontend client is generated from this spec.",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "Error": {
        "additionalProperties": true,
        "type": "object",
        "required": [
          "statusCode",
          "error",
          "message"
        ],
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Dashboard": {
        "type": "object",
        "required": [
          "key",
          "title"
        ],
        "properties": {
          "key": {
            "description": "Capability token — open /d/{key} to restore this dashboard",
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateDashboardBody": {
        "type": "object",
        "properties": {
          "title": {
            "minLength": 1,
            "maxLength": 120,
            "description": "Optional title (defaults to \"My Dashboard\")",
            "type": "string"
          }
        }
      },
      "WidgetType": {
        "type": "string",
        "enum": [
          "line",
          "bar",
          "text"
        ],
        "description": "line | bar | text"
      },
      "Period": {
        "type": "string",
        "enum": [
          "day",
          "week",
          "month",
          "year"
        ],
        "description": "day | week | month | year"
      },
      "Widget": {
        "type": "object",
        "required": [
          "id",
          "type",
          "rank",
          "title",
          "text",
          "period"
        ],
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/WidgetType"
          },
          "rank": {
            "description": "Fractional order key; widgets sort by this ascending",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "text": {
            "description": "Body of a text widget; null for charts",
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          }
        }
      },
      "WidgetPage": {
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Widget"
            }
          },
          "total": {
            "description": "Total widgets on the dashboard (across all pages)",
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      },
      "CreateWidgetBody": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/WidgetType"
          },
          "title": {
            "minLength": 1,
            "maxLength": 120,
            "type": "string"
          },
          "text": {
            "maxLength": 5000,
            "description": "Initial text (text widgets only)",
            "type": "string"
          }
        }
      },
      "UpdateWidgetBody": {
        "type": "object",
        "properties": {
          "title": {
            "minLength": 1,
            "maxLength": 120,
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "period": {
            "$ref": "#/components/schemas/Period"
          }
        }
      },
      "ReorderBody": {
        "type": "object",
        "required": [
          "orderedIds"
        ],
        "properties": {
          "orderedIds": {
            "description": "Widget ids in the desired grid order",
            "type": "array",
            "items": {
              "format": "uuid",
              "type": "string"
            }
          }
        }
      },
      "MoveWidgetBody": {
        "type": "object",
        "required": [
          "position"
        ],
        "properties": {
          "position": {
            "minimum": 0,
            "description": "Target index in the ordered list",
            "type": "integer"
          }
        }
      },
      "ChartData": {
        "type": "object",
        "required": [
          "period",
          "points"
        ],
        "properties": {
          "period": {
            "$ref": "#/components/schemas/Period"
          },
          "points": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "label",
                "positive",
                "neutral",
                "negative"
              ],
              "properties": {
                "label": {
                  "type": "string"
                },
                "positive": {
                  "type": "integer"
                },
                "neutral": {
                  "type": "integer"
                },
                "negative": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
    "/api/dashboards": {
      "post": {
        "operationId": "createDashboard",
        "summary": "Create a new anonymous dashboard",
        "tags": [
          "dashboards"
        ],
        "description": "Returns a fresh capability key. Persist it to restore the dashboard later.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDashboardBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{key}": {
      "get": {
        "operationId": "getDashboard",
        "summary": "Fetch a dashboard by key",
        "tags": [
          "dashboards"
        ],
        "description": "Used by the frontend to validate a pasted key and restore a saved dashboard.",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{key}/widgets": {
      "get": {
        "operationId": "listWidgets",
        "summary": "List a dashboard’s widgets, paged and ordered by rank",
        "tags": [
          "widgets"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetPage"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWidget",
        "summary": "Add a widget to a dashboard",
        "tags": [
          "widgets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWidgetBody"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{key}/widgets/reorder": {
      "put": {
        "operationId": "reorderWidgets",
        "summary": "Reorder a dashboard’s widgets",
        "tags": [
          "widgets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderBody"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Widget"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{key}/widgets/{id}/position": {
      "put": {
        "operationId": "moveWidget",
        "summary": "Move a widget to a target index in the dashboard order",
        "tags": [
          "widgets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveWidgetBody"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Widget id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{key}/widgets/{id}": {
      "patch": {
        "operationId": "updateWidget",
        "summary": "Update a widget (title, text, or position)",
        "tags": [
          "widgets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWidgetBody"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Widget id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteWidget",
        "summary": "Delete a widget",
        "tags": [
          "widgets"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Widget id"
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{key}/widgets/{id}/data": {
      "get": {
        "operationId": "getWidgetData",
        "summary": "Deterministic chart data for a chart widget",
        "tags": [
          "widgets"
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/Period"
            },
            "in": "query",
            "name": "period",
            "required": false
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Widget id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartData"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/{key}/widgets/{id}/regenerate": {
      "post": {
        "operationId": "regenerateWidget",
        "summary": "Reseed a chart widget (new random data)",
        "tags": [
          "widgets"
        ],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "key",
            "required": true,
            "description": "Dashboard key"
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Widget id"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Widget"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "dashboards",
      "description": "Anonymous dashboard sessions (create + restore by key)"
    },
    {
      "name": "widgets",
      "description": "Widgets scoped to a dashboard"
    }
  ]
}