# fe/src/lib/api/generated/model/moveWidgetBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * YouScan Dashboard API
 * Anonymous widget dashboards — create a dashboard, add line/bar/text widgets, restore later by key. The frontend client is generated from this spec.
 * OpenAPI spec version: 1.0.0
 */

export interface MoveWidgetBody {
  /**
   * Target index in the ordered list
   * @minimum 0
   */
  position: number;
}