# fe/src/lib/api/generated/model/widget.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * YouScan Dashboard API
 * Anonymous widget dashboards — create a dashboard, add line/bar/text widgets, restore later by key. The frontend client is generated from this spec.
 * OpenAPI spec version: 1.0.0
 */
import type { WidgetType } from './widgetType';
import type { WidgetText } from './widgetText';
import type { Period } from './period';

export interface Widget {
  id: string;
  type: WidgetType;
  /** Fractional order key; widgets sort by this ascending */
  rank: string;
  title: string;
  /** Body of a text widget; null for charts */
  text: WidgetText;
  period: Period;
}