# fe/src/lib/api/generated/model/widgetPage.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * YouScan Dashboard API
 * Anonymous widget dashboards — create a dashboard, add line/bar/text widgets, restore later by key. The frontend client is generated from this spec.
 * OpenAPI spec version: 1.0.0
 */
import type { Widget } from './widget';

export interface WidgetPage {
  items: Widget[];
  /** Total widgets on the dashboard (across all pages) */
  total: number;
  offset: number;
  limit: number;
}